% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{GCV_HuberQp}
\alias{GCV_HuberQp}
\title{Cross-validation for Huber Regression}
\usage{
GCV_HuberQp(
  lambda,
  Z,
  Y,
  H,
  vrs = "C",
  tuning = NULL,
  custfun = NULL,
  OSQP_res_abs = 1e-06,
  OSQP_res_rel = 1e-06
)
}
\arguments{
\item{lambda}{A candidate parameter value; non-negative real number.}

\item{Z}{Data matrix of dimension \code{n}-times-\code{p}, where \code{n} is
the number of observations, \code{p} is the dimension.}

\item{Y}{Vector of responses of length \code{n}.}

\item{H}{Penalty matrix of size \code{p}-times-\code{p} that
is used inside the quadratic term for penalizing estimated parameters.}

\item{vrs}{Version of the algorhitm to be used in function \link{HuberQp}; 
either \code{vrs="C"} for the \code{C++} version, or \code{vrs="R"} for the 
\code{R} version. Both should give (nearly) identical results, see 
\link{HuberQp}.}

\item{tuning}{A non-negative tuning constant for the Huber loss 
function. If left to NULL, defaults to 1.345.}

\item{custfun}{A custom function combining the residuals \code{resids}, the 
hat values \code{hats}. The result of the function must be numeric, see
\link{GCV_crit}.}

\item{OSQP_res_abs}{Absolute tolerance for OSQP quadratic programming solver (default: 1e-6).}

\item{OSQP_res_rel}{Relative tolerance for OSQP quadratic programming solver (default: 1e-6).}
}
\value{
A named numerical vector of values. The length of the vector depends
on the input. The vector contains the values:
\itemize{
 \item{"AIC"}{ Akaike's information criterion given by 
 \code{mean(resids^2)+log(n)*mean(hats)}, where \code{n} is the length of
 both \code{resids} and \code{hats}.}
 \item{"GCV"}{ Leave-one-out cross-validation criterion given by
 \code{mean((resids^2)/((1-hats)^2))}.}
 \item{"GCV(tr)"}{ Modified leave-one-out cross-validation criterion 
 given by \code{mean((resids^2)/((1-mean(hats))^2))}.}
 \item{"BIC"}{ Bayes information criterion given by 
 \code{mean(resids^2)+2*mean(hats)}.}
 \item{"rGCV"}{ A robust version of \code{GCV} where mean is replaced
 by a robust M-estimator of scale of \code{resids/(1-hats)}, see 
 \link[robustbase]{scaleTau2} for details.}
 \item{"rGCV(tr)"}{ Modified version of a \code{rGCV} given by 
 a robust M-estimator of scale of \code{resids/(1-mean(hats))}.}
 \item{"custom"}{ The custom criterion given by function \code{custfun}. 
 Works only if \code{custfun} is part of the input.}
}
}
\description{
Provides the cross-validation indices from \link{GCV_crit} in
conjunction with the \link{HuberQp} function directly as an argument of the
parameter \code{lambda}.
}
\details{
Function \code{custfun} has two arguments 
corresponding to \code{resids} and \code{hats}. The output of the function 
must be numeric. Additional parameters passed to \code{HuberQp} may be passed
to the function.
}
\examples{
n = 50      # sample size
p = 10      # dimension of predictors
Z = matrix(rnorm(n*p),ncol=p) # design matrix
Y = Z[,1]   # response vector
lambda = 1  # tuning parameter for penalization
H = diag(p) # penalty matrix
    
# Run with the ridge function
res = ridge(Z, Y, lambda, H)
with(res,GCV_crit(resids,hat_values))
with(res,GCV_crit(resids,hat_values,custfun = function(r,h) 
    sum((r/(1-h))^2)))
    
GCV_ridge(lambda,Z,Y,H,custfun = function(r,h) sum((r/(1-h))^2))
}
