% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{qr_fpca_piecewise_fine}
\alias{qr_fpca_piecewise_fine}
\title{Quantile regression based on FPCA of Kato (2012)}
\usage{
qr_fpca_piecewise_fine(X, Y, t, tau = 0.5, K_max = 10, n_fine = 200)
}
\arguments{
\item{X}{Matrix of observed values of \code{X} of size 
\code{n}-times-{p}, one row per observation, columns corresponding to the 
positions in the rows of \code{t}.}

\item{Y}{Vector of responses of length \code{n}.}

\item{t}{Domain locations for the observed points of \code{X}. A numerical
vector of length \code{p}.}

\item{tau}{Order of the quantile to be used in the estimator. By default, 
\code{tau=0.5}, which corresponds to the regression median.}

\item{n_fine}{Number of equi-spaced points in the domain for the complete
function. By default, set to \code{n_fine=200}.}

\item{Kmax}{Maximum number of functional principal components to be 
considered. By default, set to \code{Kmax=10}.}
}
\value{
A list composed of
\itemize{
 \item{"fitted_Y"}{ A vector of \code{n} fitted values.}
 \item{"alpha_hat"}{ Estimate of \code{alpha0}, a numerical value.}
 \item{"beta_hat"}{ Estimate of the regression function \code{beta0} 
 evaluated at the \code{p} points from \code{tobs}, where \code{X} was
 observed.}
 \item{"gcv_vals"}{ Observed values of the generalized cross-validation 
 criteria for the selection of the dimension. A numerical vector of length
 \code{Kmax}.}
 \item{"best_m"}{ Selected dimension using generalized cross-validation.}
}
}
\description{
Fits a (robust) quantile regression estimator in a scalar-on-function 
regression problem with discretely observed predictors. The estimator is 
based on quantile regression applied in the context of functional principal
components. The tuning parameter \code{J} (the number of principal 
compontents chosen) is selected using cross-validation.
}
\examples{
n = 50      # sample size
p = 10      # dimension of predictors
X = matrix(rnorm(n*p),ncol=p) # design matrix
Y = X[,1]   # response vector
tobs = sort(runif(p)) 

(res = qr_fpca_piecewise_fine(X, Y, tobs))
}
\references{
Kato, K. (2012). Estimation in functional linear quantile regression. 
\emph{The Annals of Statistics}, 40(6), 3108-3136.
}
