% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rho}
\alias{rho}
\title{Loss functions}
\usage{
rho(t, type, Hk = 1.345)
}
\arguments{
\item{t}{Vector of input values of length \code{n}.}

\item{type}{Integer code for the type of loss function. Accepted are
\code{type="absolute"} for the absolute loss \code{rho(t)=|t|}; 
\code{type="square"} for the square loss \code{rho(t)=t^2}; 
\code{type="Huber"} for the Huber loss \code{rho(t)=t^2/2} if 
\code{|t|<Hk} and \code{rho(t)=Hk*(|t|-Hk/2)} otherwise; and 
\code{type="logistic"} for the logistic loss 
\code{rho(t)=2*t + 4*log(1+exp(-t))-4*log(2)}.}

\item{Hk}{Tuning parameter, a non-negative real number, affects only
the Huber loss (by default \code{Hk = 1.345}.}
}
\value{
A numerical vector of values.
}
\description{
Several typically used loss functions \code{rho}.
}
\examples{
curve(rho(x,type="absolute"),-5,5) # absolute loss with tuning
curve(rho(x,type="square"),-5,5) # square loss
curve(rho(x,type="Huber"),-5,5) # Huber loss
curve(rho(x,type="logistic"),-5,5) # logistic loss
}
