% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solvers.R
\name{vorArea}
\alias{vorArea}
\title{Area of the cells in Voronoi tesselation}
\usage{
vorArea(x, I.method = "chull", I = NULL, scale = TRUE, plot = FALSE)
}
\arguments{
\item{x}{The dataset whose Voronoi tesselation should be considered.
\code{p}-times-\code{d} matrix, one row per point in 
\code{d}-dimensional space. If \code{d==1}, the elements of \code{x} must
be sorted.}

\item{I.method}{Input method for the complete domain \code{I}
where the Voronoi tesselation is evaluated. Takes a value \code{"box"}
for \code{I} the smallest axes-aligned box, or \code{"chull"}
for \code{I} being a convex hull of points. By default set to 
\code{"chull"}. In dimension \code{d=1}, the two methods \code{"box"}
and \code{"chull"} are equivalent.}

\item{I}{A set of points specifying the complete domain \code{I}.
In general, can be a \code{q}-times-\code{d} matrix, where \code{q}
is the number of points and \code{d} is the dimension. The matrix \code{I}
then specifies the point from which to compute the domain \code{I}: 
(a) If \code{method.I=="chull"}, the domain is the convex hull of \code{I}; 
(b) If \code{method.I=="box"}, the domain is the smallest axes-aligned box
that contains \code{I}. If \code{I} is \code{NULL} (by default), then 
\code{I} is taken to be the same as \code{x}. If \code{I.method=="box"}, 
\code{I} can be specified also by a pair of real values \code{a<b}, 
in which case we take \code{I} to be the axis-aligned sqare \code{[a,b]^d}.}

\item{scale}{Should the resulting volumes be scaled to sum to \code{1}? By
default set to \code{TRUE}.}

\item{plot}{A logical indicator of whether the resulting Voronoi cells should
be plotted. By default \code{FALSE}.}
}
\value{
A numerical vector of length \code{p} of volumes of the Voronoi cells
corresponding to the rows of \code{x}. If \code{scale==TRUE} (default), this
vector is scaled so that the sum of its elements is \code{1}. If some of the
elements of this vector is numerically zero (that is, some of the Voronoi 
cells are empty), a warning is given.
}
\description{
For one-dimensional or two-dimensional sets of points, 
finds the Voronoi tesselation of the data and returns
a vector of volumes (lengths or areas) of all the Voronoi 
cells.
}
\examples{
p = 50      # number of observed points in domain
x = runif(p) # x-coordinates of the points
y = runif(p) # y-coordinates of the points

# Voronoi areas in the plane:
# (a) I is the smallest box containing the data
vorArea(cbind(x,y), I.method="box",plot=TRUE)

# (b) I is the box [0,1]^2
vorArea(cbind(x,y),I.method="box",I=c(0,1),plot=TRUE)

# (c) I is the convex hull of the data
vorArea(cbind(x,y),I.method="chull",plot=TRUE)

# Voronoi areas in the real line:
# (a) I is the smallest interval containing the data
vorArea(x = matrix(sort(x),ncol=1),I.method="chull",plot=TRUE)
# (b) I = [0,1]
vorArea(x = matrix(sort(x),ncol=1),I.method="box",I=c(0,1),plot=TRUE)
# (c) I = [-1,1]
vorArea(x = matrix(sort(x),ncol=1),I.method="chull",I=matrix(c(-1,1),ncol=1),plot=TRUE)

}
