% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solvers.R
\name{eta}
\alias{eta}
\title{Radial functions for thin-plate splines}
\usage{
eta(x, d, m)
}
\arguments{
\item{x}{Vector of input values of length \code{n}.}

\item{d}{Dimension of the domain, positive integer.}

\item{m}{Order of the spline, positive integer.}
}
\value{
A numerical vector of values.
}
\description{
The function \code{eta_{m,d}} involved in the definition of a 
\code{d}-dimensional thin-plate spline of order \code{m}.
}
\examples{
curve(eta(x,d=1,m=3),-5,5) 

}
