% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solvers.R
\name{rho}
\alias{rho}
\title{Loss functions}
\usage{
rho(t, type, alpha = 1/2, tuning = 1.345)
}
\arguments{
\item{t}{Vector of input values of length \code{n}.}

\item{type}{The type of the loss function used in the minimization problem.
Accepted are \code{type="absolute"} for the absolute loss \code{rho(t)=|t|/2};
\code{type="quantile"} for the (asymmetric) quantile loss 
\code{rho(t)=t(alpha-I[t<0])} (\code{absolute} loss with \code{alpha=1/2});
\code{type="square"} for the square loss \code{rho(t)=t^2}; 
\code{type="Huber"} for the Huber loss \code{rho(t)=t^2/2} if 
 \code{|t|<tuning} and \code{rho(t)=tuning*(|t|-tuning/2)} otherwise; and
\code{type="logistic"} for the logistic loss 
\code{rho(t)=2*t + 4*log(1+exp(-t))-4*log(2)}.}

\item{alpha}{The order of the quantile if \code{type="quantile"}. By default
taken to be \code{alpha=1/2}, which gives the absolute loss 
(\code{type="absolute"}).}

\item{tuning}{tuning parameter, a non-negative real number, affects only
the Huber loss (by default \code{tuning = 1.345}.}
}
\value{
A numerical vector of values.
}
\description{
Several typically used loss functions \code{rho}.
}
\examples{
curve(rho(x,type="absolute"),-5,5) # absolute loss with tuning
curve(rho(x,type="quantile", alpha = 0.1),-5,5) # quantile loss 
curve(rho(x,type="square"),-5,5) # square loss
curve(rho(x,type="Huber"),-5,5) # Huber loss
curve(rho(x,type="logistic"),-5,5) # logistic loss

}
